Re consolidation of MiiSsample based on env varaiable
